Extract out the custom datetime formatting to utils
